requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-requests
robotframework-jsonlibrary
robotframework-datadriver
robotframework-faker
